move validate_inputs to util as well
