+++ particula/util/tests/constants_test.py
@@ -1,8 +1,9 @@
 """ testing getting constants from the get_constants.py file
 """
 
-from particula.util.constants import (AVOGADRO_NUMBER, BOLTZMANN_CONSTANT,
-                                 GAS_CONSTANT)
+from particula.util.constants import (AVOGADRO_NUMBER,
+                                      BOLTZMANN_CONSTANT,
+                                      GAS_CONSTANT)
 
 
 def test_constants():
