+++ particula/util/validate_inputs.py
@@ -3,6 +3,7 @@
 
 import inspect
 from functools import wraps
+
 import numpy as np
